pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Filip-Povidernyi <p.poviderniy@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"
fastapi = { version = "^0.115.12", extras = ["standard"] }
uvicorn = { version = "^0.34.2", extras = ["standard"] }
pydantic = "^2.11.5"
python-jose = { version = "^3.5.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
libgravatar = "^1.0.4"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
slowapi = "^0.1.9"
redis = "^6.2.0"
fastapi-mail = "^1.5.0"
cloudinary = "^1.44.1"

[tool.poetry.group.dev.dependencies]
faker = "^37.3.0"
asyncio = "^3.4.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
